requirements.txt
absl-py==0.12.0
albumentations @ file:///home/conda/feedstock_root/build_artifacts/albumentations_1606664784723/work
astor==0.8.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1616843134209/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
cached-property==1.5.2
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246439577/work
chardet==4.0.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
config==0.5.0.post0
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
decorator==4.4.2
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
docopt==0.6.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
et-xmlfile==1.0.1
gast==0.4.0
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imgaug==0.4.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1616844030656/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1613852943625/work/dist/ipykernel-5.5.0-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1612119311452/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1612125253553/work
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.4
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1602267314491/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1597876339545/work
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1605115652956/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1611005694671/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1594379821643/work
numpy==1.19.5
olefile==0.46
openpyxl==3.0.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandocfilters==1.4.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///tmp/build/80754af9/pillow_1602788709045/work
pipreqs==0.4.10
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit==1.0.15
protobuf==3.15.6
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1615243893546/work
pyparsing==2.4.7
pyrsistent @ file:///home/conda/feedstock_root/build_artifacts/pyrsistent_1605115583744/work
python-dateutil==2.8.1
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658315997/work
PyYAML==5.4.1
pyzmq==19.0.2
requests==2.25.1
scikit-image==0.17.2
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
Send2Trash==1.5.0
Shapely @ file:///home/conda/feedstock_root/build_artifacts/shapely_1611315601052/work
simplegeneric==0.8.1
six==1.15.0
tensorboard==1.12.2
tensorboard-logger==0.1.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1616072119018/work
testpath==0.4.4
tifffile==2020.10.1
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.0.0
torchvision==0.2.1
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
traitlets==4.3.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3==1.26.6
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
